+++ src/mqtt.py
@@ -243,6 +243,8 @@ def update_car_data(force_update=False, overwrite={}):
                 state = last_data_update
             elif entity["id"] == "active_schedules":
                 state = active_schedules[vin]
+            elif entity["id"] == "api_backend_status":
+                state = volvo.backend_status
             else:
                 if entity["id"] == ov_entity_id and vin == ov_vin:
                     state = ov_state